replaced if/else function with switch function. Removed icon reference object list. created a colour changing function.
To try and speed up the page’s load time, I replaced if/else function
with a switch function. I removed the iconIndex object list because
there was no need to use it. I attached comments for reference to the
switch function instead. The iconIndex used up too much space.

I stored a Color changing function into a variable so it could just be
reused. The switch function reuses this function for each ID.

Prior to these changes, there was a lot of repetitive code and it was
all mixed and confusing. The project could still be made more
efficient, with the use of one or two more functions. The switch
function is still quite repetitive. I may also use clearer
understandable names for variables and objects.