Changed project's Javascript and HTML
The project now only detects weather from Sydney, Australia because of
security issues using the ip api. I have kept the removed javascript in
a tag, for reference and potential future use. I also added the
#lotData html to the index.html file directly because there is no need
to append it anymore. I am planning to make the project iterate through
the html dom, instead of just appending the data with HTML tags. This
is to keep the code cleaner and because it’s better practice.